pyproject.toml
[project]
name = "octpus-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "aiohttp>=3.12.15",
  "faiss-cpu>=1.11.0.post1",
  "fastapi>=0.116.1",
  "ipython>=9.4.0",
  "langchain-community>=0.3.27",
  "langchain-core>=0.3.72",
  "langchain-deepseek>=0.1.4",
  "langchain-mcp-adapters>=0.1.9",
  "langchain-tavily>=0.2.11",
  "langgraph>=0.6.1",
  "mcp[cli]>=1.12.2",
  "nest-asyncio>=1.6.0",
  "pydantic>=2.11.7",
  "python-dotenv>=1.1.1",
  "uvicorn>=0.35.0",
]
